# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Karag API",
    "description": "Modular RAG & Agentic Chatbot API",
    "contact": {
      "name": "Karag Team",
      "url": "https://github.com/karag-ai/karag"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/chat/history/{thread_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat History",
        "description": "Get chat history for a specific thread.",
        "operationId": "get_chat_history_api_v1_workspaces__workspace_id__chat_history__thread_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_list_ChatMessage__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/chat/threads": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Chat Threads",
        "description": "List chat threads with pagination support.",
        "operationId": "list_chat_threads_api_v1_workspaces__workspace_id__chat_threads_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "examples": [
                1
              ],
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (max 100)",
              "examples": [
                20
              ],
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_list_ThreadMetadata__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/chat/threads/{thread_id}/title": {
      "patch": {
        "tags": [
          "chat"
        ],
        "summary": "Update Thread Title",
        "description": "Update the title of a chat thread.\n\nArgs:\n    thread_id: The unique identifier of the chat thread\n    payload: Contains the new title\n    workspace: Current workspace context\n\nReturns:\n    AppResponse confirming the update",
        "operationId": "update_thread_title_api_v1_workspaces__workspace_id__chat_threads__thread_id__title_patch",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadTitleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/chat/threads/{thread_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Thread",
        "description": "Get a specific chat thread by ID.",
        "operationId": "get_thread_api_v1_workspaces__workspace_id__chat_threads__thread_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_ThreadMetadata_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete Thread",
        "description": "Delete a chat thread.\n\nArgs:\n    thread_id: The unique identifier of the chat thread\n    workspace: Current workspace context\n\nReturns:\n    AppResponse confirming the deletion",
        "operationId": "delete_thread_api_v1_workspaces__workspace_id__chat_threads__thread_id__delete",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Stream",
        "description": "Stream chat updates using Server-Sent Events (SSE).\n\nThis endpoint streams chat responses in real-time and spawns\nbackground title generation for a consistent UX.\n\nArgs:\n    payload: Chat stream request containing message and configuration\n    workspace: Current workspace context\n\nReturns:\n    StreamingResponse with text/event-stream content type\n\nExample:\n    Client connects and receives SSE events:\n    - data: {\"type\": \"token\", \"content\": \"Hello\"}\n    - data: {\"type\": \"token\", \"content\": \" world\"}\n    - data: {\"type\": \"done\"}",
        "operationId": "chat_stream_api_v1_workspaces__workspace_id__chat_stream_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse",
        "description": "Standardized API response wrapper.\n\nAll API responses follow this structure for consistency:\n{\n    \"success\": bool,\n    \"code\": str,\n    \"message\": str,\n    \"data\": T | null\n}\n\nType Parameters:\n    T: The type of data contained in the response\n\nExamples:\n    >>> # Success response with data\n    >>> response = AppResponse.success_response(\n    ...     data={\"id\": 1, \"name\": \"Test\"},\n    ...     message=\"User created successfully\"\n    ... )\n\n    >>> # Error response\n    >>> response = AppResponse.business_failure(\n    ...     code=\"VALIDATION_ERROR\",\n    ...     message=\"Invalid input provided\"\n    ... )"
      },
...
      "AppResponse_ThreadMetadata_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThreadMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[ThreadMetadata]"
      },
...
      "AppResponse_list_ChatMessage__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[list[ChatMessage]]"
      },
...
      "AppResponse_list_ThreadMetadata__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ThreadMetadata"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[list[ThreadMetadata]]"
      },
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "The role of the message sender (user or assistant)"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "The content of the message"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Unique identifier for the message"
          },
          "reasoning_steps": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Steps",
            "description": "Thinking process steps"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources",
            "description": "Source documents used for the response"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
      "ChatStreamRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "The user message to process"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "The ID of the chat thread"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "description": "The workspace ID context"
          },
          "execution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution",
            "description": "Runtime execution settings"
          }
        },
        "type": "object",
        "required": [
          "message",
          "thread_id",
          "workspace_id"
        ],
        "title": "ChatStreamRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ThreadMetadata": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": "New chat"
          },
          "has_thinking": {
            "type": "boolean",
            "title": "Has Thinking",
            "default": false
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "workspace_id"
        ],
        "title": "ThreadMetadata"
      },
      "ThreadTitleUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ThreadTitleUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}